requirements.txt
tqdm==4.66.1
wheel
pycocotools==2.0.7
scikit-learn>=1.4.0
requests
inflect==7.0.0
fiftyone==0.24.0
fairscale==0.4.4
sentencepiece==0.1.99
modelscope==1.9.0
accelerate>=0.24.0
diffusers>=0.27.0
huggingface_hub>=0.20.2,<1.0
click==8.1.7
open_clip_torch
openai-clip
promptcap==1.0.3
# salesforce-lavis==1.0.2  # Requires decord which needs to be built from source on macOS ARM64
transformers>=4.35.0
ninja==1.11.1.1
